+++ NEWS.md
@@ -1,4 +1,4 @@
-# ggside 0.3.2.9999
+# ggside 0.4.0
 
 ## Details/Breaking Changes
 
@@ -10,7 +10,7 @@ Below is a running log of changes that are required due to `ggplot2`'s welcome c
   * `ggside` no longer exports its own overload of `+.gg`.
   * `ggside` now uses the recommened generic `ggplot2::update_ggplot()` to control how `ggside` objects are added to this object.
   * previously, `ggside` would rely on `NextMethod()` to flexibly dispatch on other `ggplot2` subclasses. With `S7` this is no longer easily done. For example, `update_ggplot(object = <class_any>, plot = <class_ggside>)` explicitly forces the `plot` argument to be dispatched as a `class_ggplot` object. This prevents possible multiple inheritance interactions with `ggside` and other extensions of `ggplot2`.
-* `ggside` (> v0.3.2) will now depend on `ggplot2 (> v3.5.2.xxx)`.
+* `ggside` (>= v0.4.0) will now depend on `ggplot2 (>= v4.0.0)`.
 
 # ggside 0.3.2